LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/open-during-abort-processing-expected.txt


FAIL XMLHttpRequest: open() during abort processing - abort() called from onloadstart assert_array_equals: property 7, expected "readyState after abort() 1" but got "readyState after abort() 0"
